requirements.txt
tensorflow
numpy
matplotlib
seaborn
streamlit
pillow